Still debuggin pit. From the bochs logs, it seems like I'm writing 0, when I should be writing 1 to lobyte.
In the meantime, I added a link time string to be added on each build because I had been running the wrong build for some time and didn't know it. I do still have a problem with file dependencies not being tracked properly; and therefore not being correct. This is a start though.

I want to get a PIT interrupt working next, still.