Have a timer mechanism and delay working
I don't know if the math for the timer is correct. I tried a 2 second delay, and it took approximately 10 seconds, but I don't  know if this is because of bochs, or a bug in my math (I believe the latter).

I've also noticed that really short timer ticks cause spurious interrupts on the 8259a. Again, not sure of the cause of this yet.