I fixed a bug in the 8254 programming.
It seems from reading the bochs documentation that by default bochs will not
try to keep the 8254 ticking in real time. By my estimate on my machine it
is 3x slower than real time. For now I will leave this as is, because it seems
that the programming is indeed correct.

I should probably triple check it later.