Various changes, nothing substantial
Embedded the block device structure in the ATA structure. I meant to do
this earlier, but am only getting around to it now. This had an effect
on the ATA files.

Also a couple of partition related changes which I didn't feel like
pulling out for this commit.

Finally a lingering Makefile fix from the last round.