smallboot/ext2
Some new oddities seen in ext2 files on recent builds. I was seeing in the
doubly indirect blocks pointers to zero. Followed by real pointers. In
reversing this a bit, I found that a pointer to zero must in fact be zero
filled by the bootloader. I need to go back and find it in the spec because my
searching didn't find anything related to this.

Also possible bugfix in here.