common/cdefs: Clarify the mess
I was too lazy to figure out how to properly deal with CONCAT (since my
usage, similar to FreeBSD, was different than that of the GNU/Linux
usage). It wasn't ever right, but to make progress I copied the cdefs.h
from FreeBSD.

This was not necessary. This patch attempts to address this mess by
declaring my own __CAT instead of __CONCAT, and using that.

Previously I only ever used this from the bootloader, which is
freestanding, so it didn't matter. I found this when writing a tool
which wanted to include the standard libraries and these macros (elf
parser).