linker script cleanup
In preparation for the stack backtraces with symbols, I need to much
with the sections which get loaded. As the linker scripts were
atrocious, I decided to clean that up first.

UPDATE: I read a bit more about the multiboot spec, and realized that is
the proper way to transfer the symbol table. As such, this serves only
as a cleanup.