smallboot/elf: Load multiboot symbols
Passing the symbol info to the kernel is supported by the multiboot spec
(and I did correctly do this on the initial implementation). The catch
is, symbol information are in non-loadable sections [not segments], To
support this, I've made the bootloader update the addr field in the
section header, so that the kernel can read out the relevant info. This
seems to be how GRUB works are well.

While here, clean up the code.