remove early_printf
Consolidate the print code by only requiring early_putc. In addition
actually get the ringbuffer early print code working. The code is
accessible as a global variable.

This now deprecates the need for the early printf config option, and
instead we can provide the port E9 config option

As a bonus, this makes assertions before the console is up not die in a
fire.