common/builtins: Start a builtins header
This is meant for builtins which would normally be used by the compiler.
I am just not sure if it's safe to use __builtin_* in many cases since I
am not sure what library function it may inevitably depend on. I would
hope that since the compile is invoked with -nostdlib, it will dtrt. I
can try it later.