buildsystem Cleanups
This addresses the new warning generated as a result of the use of
imacros
commit 1ec4cea579f72244e8cc86a9a7701c52cc796d4a
Author: Ben Widawsky <ben@bwidawsk.net>
Date:   Sun Oct 26 14:38:31 2014 -0700

    cdefs: Move OFFSET_OF to cdefs.

makedepend is unable to recognize that flag, and therefore that commit
broke proper dependency tracking of cdefs.h. For now I will keep imacros
because cdefs.h doesn't change very often. There is a new variable
defined called ESOTERIC_INCLUDES (can be named ESOTERIC_CFLAGS later)
which will be ignored by makedepend.

While in the build goop, I added some cleanups and comments.

Also while in the goop, fix up warnings. By design, Make will spit out a
warning if the operation failed, even if the command is preceded with a
'-'. I find this annoying.