includes: Try to fix the include situation somewhat
All standard includes should use the "<" ">" for includes. Anything with quotes
("foo.h") is either in need of fixing via  a proper interface, or is a relative
path to the current directory.

I don't claim this to be the entire fix and that everything is now in good
shape, but it's a start.