kern/Makefile.base: use QUIETFLAGS instead of MAKEFLAGS
For some reason that I cannot explain, and do not care to debug, MAKEFLAGS was
being set to "w" when make was being invoked with -C from another directory, ie.
make -C kern.

Since I want to be able to have the top level make the kernel, this needed to be
fixed.